requirements.txt
djangorestframework-jsonp
djangorestframework-jwt
certifi==2017.7.27.1
chardet==3.0.4
codegen==1.0
configparser==3.5.0
Django==2.0.6
django-bootstrap-form==3.2.1
django-pagination==1.0.7
djangorestframework==3.8.2
horas==0.0.1a4
idna==2.5
Markdown==2.6.11
Naked==0.1.31
patterns==0.3
psycopg2==2.7.1
pytz==2017.2
PyYAML==3.12
requests==2.18.3
status==0.2.5
urllib3==1.22
whitenoise==3.3.0
xlwt==1.3.0
django-grappelli